feat(radar): extract FilterPanel + add multi-select Roles section
P16 Phase 3. Reality-check on the prior plan: FilterPanel was inline
in HotZoneLayer.tsx (~190 lines), and the filter shape was
{showMode, operator, region}, NOT {operator, tail} as the original
P11 plan described. This PR aligns with the actual state.

Changes:

- New components/FilterPanel.tsx — inline FilterPanel + Group + Pill
  helpers extracted verbatim, then extended with a Roles section.
- New filter field: roles: FilterableRole[] (smokey | patrol | sar |
  transport | unknown). Empty array = no role constraint (back-compat
  with prior persisted state).
- Filters compose: passesAircraftFilter() applies roles allow-list
  AND existing showMode predicate (role ∈ allow-list AND showMode
  passes). Both must hold.
- buildQueryString() in HotZoneLayer.tsx forwards roles to
  /api/hotzones (server already accepts comma-separated `roles`).
- readRadarFilter() migration: missing roles[] → []. Same persisted
  storage key (ss_hotzones_filter), no localStorage migration needed.
- HotZoneLayer.tsx now imports FilterPanel from its own file and
  drops the local definitions + the now-unused OPERATORS import.

Out of scope (deferred):
- operators[] multi-select — current UI keeps single-operator under
  showMode='operator', which is sufficient for the verify-prod
  assertion. Multi-operator can land in a follow-up.
- Layers / weather toggle — that lives in Phase 4 (NX2).
